Use Promise Spec Terminology Where Appropriate
Use "Fulfilled" in place of "resolved".
Use "Pending" in place of "loading" when "loading" didn't feel more natural.
Capitalize "O" in object when used as a datatype to use capitalization preferred by JSDoc and other tools. Only `boolean`, `number`, and `string` are expected to be all lowercase (assuming you are expecting the primitives and not the wrapper objects).
Fix a few minor mistakes (grammar, unescaped tag syntax, etc).