improvement: `deferFn` and `promiseFn` now have the same signature.
The `promiseFn` and the `deferFn` have been unified. They now share the
following signature:

```ts
export type AsyncFn<T, C> = (
  context: C | undefined,
  props: AsyncProps<T, C>,
  controller: AbortController
) => Promise<T>
```

Before the `deferFn` and `promiseFn` had this signature:

```ts
export type PromiseFn<T> = (props: AsyncProps<T>, controller: AbortController) => Promise<T>

export type DeferFn<T> = (
  args: any[],
  props: AsyncProps<T>,
  controller: AbortController
) => Promise<T>
```

The big change is the introduction of the `context` parameter. The idea
behind this parameter is that it will contain the parameters which are
not known to `AsyncOptions` for use in the `promiseFn` and `asyncFn`.

Another goal of this commit is to make TypeScript more understanding
of the `context` which `AsyncProps` implicitly carries around. Before
this commit the `AsyncProps` accepted extra prop via `[prop: string]: any`.
This breaks TypeScript's understanding of the divisions somewhat. This
also led to missing types for `onCancel` and `suspense`, which have been
added in this commit.

To solve this we no longer allow random extra properties that are unknown
to `AsyncProps`. Instead only the new `context` of `AsyncProps` is passed.
This means that the `[prop: string]: any` of `AsyncProps` is removed this
makes TypeScript understand the props better.

The other big change of this commit that `useAsync` no longer supports
an overload. This means that the user can no longer do:

```ts
const state = useAsync(loadPlayer, { context: { playerId: 1 } })
```

But have to be more explicit:

```t
const state = useAsync({ promiseFn: loadPlayer, context: { playerId: 1 } })
```

These changes are of course a breaking change.

Also now compiling TypeScript on `yarn test` this should prevent type
errors from slipping in.

Closes: #246

WIP: Trying to fix build

asdf