Added an explicit browserlist for Parcel builds
Added a browserlist to indicate support for modern ES. When Parcel detects a library's browserlist is different then the application browser list it will transpile the library to the version of ES required by the application. This allows developers to specify older browsers (eg. ie 11) and have react-async transparently transpiled.